drm/amdgpu: avoid KIQ clashing with compute or KFD queues
Instead of picking an arbitrary queue for KIQ, search for one according
to policy. The queue must be unused.

Also report the KIQ as an unavailable resource to KFD.

In testing I ran into KCQ initialization issues when using pipes 2/3 of
MEC2 for the KIQ. Therefore the policy disallows grabbing one of these.

Signed-off-by: Andres Rodriguez <andresx7@gmail.com>