drm/amdgpu: move priority decay logic into amd_sched_priority_ctr
So that it can be re-used.

A priority counter can be used to track priority requests. If no active
requests exists, the counter will default to ()->default_priority.

The re-factored version is now allowed to decay below NORMAL. This
allows us to create requests that lower the priority. This didn't
matter for rings, as no priority below NORMAL has an effect.

Signed-off-by: Andres Rodriguez <andresx7@gmail.com>