drm/amdgpu: add interface for editing a foreign process's priority v2
The AMDGPU_SCHED_OP_PROCESS_PRIORITY_GET/PUT ioctls are used to set the
priority of a different process in the current system.

When all requests are dropped, the foreign process's contexts will be
restored to the priority specified at context creation time.

An fd is used to identify the remote process. This is simpler than
passing a pid number, which is vulnerable to re-use, etc.

This functionality is limited to DRM_MASTER since abuse of this
interface can have a negative impact on the system's performance.

v2: removed unused output structure

Signed-off-by: Andres Rodriguez <andresx7@gmail.com>